requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
httpx
supabase
bcrypt
